Implement CardTemplatePreviewer as special Previewer case
