+++ AnkiDroid/src/main/AndroidManifest.xml
@@ -291,6 +291,11 @@
             android:name="com.ichi2.anki.Previewer"
             android:configChanges="locale"
             android:label="@string/preview_title" />
+        <activity
+            android:exported="false"
+            android:name="com.ichi2.anki.CardTemplatePreviewer"
+            android:configChanges="locale"
+            android:label="@string/preview_title" />
 
         <!-- small widget -->
         <receiver
