+++ AnkiDroid/src/main/java/com/ichi2/anki/CardTemplatePreviewer.java
@@ -1,0 +1,183 @@
+/***************************************************************************************
+ * Copyright (c) 2019 Mike Hardy <mike@mikehardy.net>                                   *
+ *                                                                                      *
+ * This program is free software; you can redistribute it and/or modify it under        *
+ * the terms of the GNU General Public License as published by the Free Software        *
+ * Foundation; either version 3 of the License, or (at your option) any later           *
+ * version.                                                                             *
+ *                                                                                      *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY      *
+ * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A      *
+ * PARTICULAR PURPOSE. See the GNU General Public License for more details.             *
+ *                                                                                      *
+ * You should have received a copy of the GNU General Public License along with         *
+ * this program.  If not, see <http://www.gnu.org/licenses/>.                           *
+ ****************************************************************************************/
+
+package com.ichi2.anki;
+
+import android.os.Bundle;
+
+import com.ichi2.libanki.Card;
+import com.ichi2.libanki.Collection;
+import com.ichi2.libanki.Models;
+import com.ichi2.libanki.Note;
+
+import org.json.JSONObject;
+
+import java.util.ArrayList;
+
+import javax.annotation.Nullable;
+
+import timber.log.Timber;
+
+/**
+ * The card template previewer intent must supply one or more cards to show and the index in the list from where
+ * to begin showing them. Special rules are applied if the list size is 1 (i.e., no scrolling
+ * buttons will be shown).
+ */
+public class CardTemplatePreviewer extends Previewer {
+    private String mEditedModelFileName = null;
+    private JSONObject mEditedModel = null;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        Timber.d("onCreate()");
+
+        Bundle parameters = savedInstanceState;
+        if (parameters == null) {
+            parameters = getIntent().getExtras();
+        }
+        mEditedModelFileName = parameters.getString(CardTemplateEditor.INTENT_MODEL_FILENAME);
+        mIndex = parameters.getInt("index");
+
+        if (mEditedModelFileName != null) {
+            Timber.d("onCreate() loading edited model from %s", mEditedModelFileName);
+            mEditedModel = CardTemplateEditor.getTempModel(mEditedModelFileName);
+        }
+
+        if (mEditedModel != null && mIndex != -1) {
+            Timber.d("onCreate() CardTemplatePreviewer started with edited model and template index, displaying blank to preview formatting");
+            mCurrentCard = getDummyCard(mEditedModel, mIndex);
+            if (mCurrentCard == null) {
+                UIUtils.showSimpleSnackbar(this, R.string.invalid_template, false);
+                finishWithoutAnimation();
+                return;
+            }
+        }
+
+        if (mCurrentCard == null || mIndex < 0) {
+            Timber.e("CardTemplatePreviewer started with empty card list or invalid index");
+            finishWithoutAnimation();
+            return;
+        }
+
+        super.onCreate(savedInstanceState);
+    }
+
+
+    @Override
+    public void onSaveInstanceState(Bundle outState) {
+        outState.putString(CardTemplateEditor.INTENT_MODEL_FILENAME, mEditedModelFileName);
+        super.onSaveInstanceState(outState);
+    }
+
+
+    @Override
+    protected void onCollectionLoaded(Collection col) {
+        if (mCurrentCard == null) {
+            mCurrentCard = new PreviewerCard(col, mCardList[mIndex]);
+        }
+        super.onCollectionLoaded(col);
+    }
+
+    @Override
+    protected Card getCard(Collection col, long cardListIndex) {
+        return new PreviewerCard(col, cardListIndex);
+    }
+
+    /** Get a dummy card */
+    protected @Nullable Card getDummyCard(JSONObject model, int ordinal) {
+        Timber.d("getDummyCard() Creating dummy note for position %s", ordinal);
+        if (model == null) {
+            return null;
+        }
+        Note n = getCol().newNote(model);
+        ArrayList<String> fieldNames = Models.fieldNames(model);
+        for (int i = 0; i < fieldNames.size(); i++) {
+            n.setField(i, fieldNames.get(i));
+        }
+        try {
+            JSONObject template = (JSONObject)model.getJSONArray("tmpls").get(ordinal);
+            PreviewerCard card = (PreviewerCard)getCol()._newCard(new PreviewerCard(getCol()), n, template, 1, false);
+            card.setNote(n);
+            return card;
+        } catch (Exception e) {
+            Timber.e("getDummyCard() unable to create card");
+        }
+        return null;
+    }
+
+
+    /** Override certain aspects of Card behavior so we may display unsaved data */
+    public class PreviewerCard extends Card {
+
+        private Note mNote;
+
+
+        public PreviewerCard(Collection col) {
+            super(col);
+        }
+
+
+        public PreviewerCard(Collection col, long id) {
+            super(col, id);
+        }
+
+
+        @Override
+        /** if we have an unsaved note saved, use it instead of a collection lookup */
+        public Note note(boolean reload) {
+            if (mNote != null) {
+                return mNote;
+            }
+            return super.note(reload);
+        }
+
+
+        @Override
+        /** if we have an unsaved note saved, use it instead of a collection lookup */
+        public Note note() {
+            if (mNote != null) {
+                return mNote;
+            }
+            return super.note();
+        }
+
+
+        /** set an unsaved note to use for rendering */
+        public void setNote(Note note) {
+            mNote = note;
+        }
+
+
+        @Override
+        /** if we have an unsaved note, never return empty */
+        public boolean isEmpty() {
+            if (mNote != null) {
+                return false;
+            }
+            return super.isEmpty();
+        }
+
+
+        @Override
+        /** Override the method that fetches the model so we can render unsaved models */
+        public JSONObject model() {
+            if (mEditedModel != null) {
+                return mEditedModel;
+            }
+            return super.model();
+        }
+    }
+}
