+++ AnkiDroid/src/test/java/com/ichi2/anki/TestCardTemplatePreviewer.java
@@ -1,0 +1,6 @@
+package com.ichi2.anki;
+
+public class TestCardTemplatePreviewer extends CardTemplatePreviewer {
+    public boolean getShowingAnswer() { return mShowingAnswer; }
+    public void disableDoubleClickPrevention() { mLastClickTime = (AbstractFlashcardViewer.DOUBLE_TAP_IGNORE_THRESHOLD * -2); }
+}
